requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
sqlalchemy
aiosqlite
alembic
psycopg2
asyncpg
psycopg
black
redis
fastapi-cache2
pillow
celery
pytest-dotenv
pytest-asyncio
ruff
bcrypt==4.0.1
PyJWT==2.9.0
passlib==1.7.4
pwdlib[argon2]
